requirements.txt
pandas==0.24.1
mysql-connector-python